Implement #80 and #81.  Fix some firmware checker glitches.

+++ octoprint_arc_welder/data/lib/c/gcode_processor_lib/utilities.cpp
@@ -218,7 +218,7 @@ std::istream& utilities::safe_get_line(std::istream& is, std::string& t)
 
 std::string utilities::center(std::string input, int width) 
 {
-	int input_width = input.length();
+	int input_width = (int)input.length();
 	int difference = width - input_width;
 	if (difference < 1)
 	{
